non-working bookmark commit; first set of major flow changes
