+++ src/lib/actions/lib/search-permission-sets.mjs
@@ -1,13 +1,13 @@
 import { DescribePermissionSetCommand, ListPermissionSetsCommand } from '@aws-sdk/client-sso-admin'
 
-const searchPermissionSets = async ({ instanceARN, policyName, ssoAdminClient }) => {
+const searchPermissionSets = async ({ identityStoreARN, policyName, ssoAdminClient }) => {
   let nextToken, permissionSetARN
   do {
-    const listPermissionSetsCommand = new ListPermissionSetsCommand({ InstanceArn : instanceARN })
+    const listPermissionSetsCommand = new ListPermissionSetsCommand({ InstanceArn : identityStoreARN })
     const listPermissionSetsResult = await ssoAdminClient.send(listPermissionSetsCommand)
     for (const testARN of listPermissionSetsResult.PermissionSets) {
       const describePermissionSetCommand = new DescribePermissionSetCommand({
-        InstanceArn      : instanceARN,
+        InstanceArn      : identityStoreARN,
         PermissionSetArn : testARN
       })
       const { Name: name } = (await ssoAdminClient.send(describePermissionSetCommand)).PermissionSet
