+++ src/lib/shared/find-identity-store.mjs
@@ -3,9 +3,9 @@ import { ListInstancesCommand, SSOAdminClient } from '@aws-sdk/client-sso-admin'
 
 import { progressLogger } from './progress-logger'
 
-const findIdentityStore = async ({ credentials, firstCheckRegion, instanceRegion, scope }) => {
-  const regionDescription = instanceRegion !== undefined
-    ? instanceRegion + ' region '
+const findIdentityStore = async ({ credentials, firstCheckRegion, identityStoreRegion, scope }) => {
+  const regionDescription = identityStoreRegion !== undefined
+    ? identityStoreRegion + ' region '
     : scope !== undefined
       ? scope + ' regions '
       : ''
@@ -25,9 +25,9 @@ const findIdentityStore = async ({ credentials, firstCheckRegion, instanceRegion
 
     let regions = listRegionsResult.Regions.map(({ RegionName }) => RegionName)
     const origRegions = [...regions]
-    if (instanceRegion !== undefined || scope !== undefined) {
-      const testFunc = instanceRegion !== undefined
-        ? (regionName) => regionName === instanceRegion
+    if (identityStoreRegion !== undefined || scope !== undefined) {
+      const testFunc = identityStoreRegion !== undefined
+        ? (regionName) => regionName === identityStoreRegion
         : scope.startsWith('!') === true
           ? (regionName) => !regionName.startsWith(scope.slice(1))
           : (regionName) => regionName.startsWith(scope)
@@ -68,10 +68,10 @@ const findIdentityStore = async ({ credentials, firstCheckRegion, instanceRegion
         progressLogger.write(' FOUND.\n')
         const instance = listInstancesResult.Instances[0]
         return {
-          id          : instance.IdentityStoreId,
-          instanceARN : instance.InstanceArn,
-          name        : instance.Name,
-          region,
+          identityStoreID          : instance.IdentityStoreId,
+          identityStoreARN : instance.InstanceArn,
+          identityStoreName   : instance.Name,
+          identityStoreRegion : region,
           ssoAdminClient,
           ssoStartURL : 'https://' + instance.IdentityStoreId + '.awsapps.com/start'
         }
